+++ lib/apiservers/engine/constants/constants.go
@@ -1,4 +1,4 @@
-// Copyright 2017 VMware, Inc. All Rights Reserved.
+// Copyright 2016-2018 VMware, Inc. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package proxy
+package constants
 
 const (
-	SwaggerSubstringEOF = "EOF"
+	DefaultVolumeDriver = "vsphere"
 )
