+++ lib/apiservers/engine/proxy/container_proxy_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package backends
+package proxy
 
 import (
 	"testing"
