+++ tests/test-cases/Group1-Docker-Commands/1-28-Docker-Secret.robot
@@ -26,21 +26,21 @@ ${fake-secret}  test
 Docker secret ls
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} secret ls
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  vSphere Integrated Containers does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker secret create
     Run  echo '${fake-secret}' > secret.file
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} secret create mysecret ./secret.file
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  vSphere Integrated Containers does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker secret inspect
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} secret inspect my_secret
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  vSphere Integrated Containers does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker secret rm
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} secret rm my_secret
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  vSphere Integrated Containers does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
