+++ tests/test-cases/Group1-Docker-Commands/1-31-Docker-Node.md
@@ -21,7 +21,7 @@ This test requires that a vSphere server is running and available
 8. Issue docker node inspect
 
 # Expected Outcome:
-* Step 2-6 should result in an error that contains does not yet support Docker Swarm
+* Step 2-6 should result in an error that contains Docker Swarm is not yet supported
 * Step 7-8 should result in an error that contains No such node
 
 # Possible Problems:
