+++ tests/test-cases/Group1-Docker-Commands/1-34-Docker-Stack.robot
@@ -25,24 +25,24 @@ Test Timeout  20 minutes
     #Should Be Equal As Integers  ${rc}  0
     #${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} stack deploy -c #./docker-compose-stack.yml proxy
     #Should Be Equal As Integers  ${rc}  1
-    #Should Contain  ${output}  does not yet support Docker Swarm
+    #Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker stack ls
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} stack ls
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker stack ps
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} stack ps test-stack
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker stack rm
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} stack rm test-stack
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker stack services
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} stack services test-stack
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
