+++ tests/test-cases/Group1-Docker-Commands/1-35-Docker-Swarm.robot
@@ -23,33 +23,33 @@ Test Timeout  20 minutes
 Docker swarm init
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm init
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker swarm join
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm join 127.0.0.1:2375
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker swarm join-token
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm join-token worker
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm join-token manager
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker swarm leave
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm leave
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker swarm unlock-key
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm unlock-key
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
 
 Docker swarm update
     ${rc}  ${output}=  Run And Return Rc And Output  docker1.13 %{VCH-PARAMS} swarm update --autolock
     Should Be Equal As Integers  ${rc}  1
-    Should Contain  ${output}  does not yet support Docker Swarm
+    Should Contain  ${output}  Docker Swarm is not yet supported
